+++ src/inmanta/execute/scheduler.py
@@ -36,7 +36,6 @@ from inmanta.ast.statements.define import (
     DefineTypeDefault,
 )
 from inmanta.ast.type import TYPES, Type
-from inmanta.config import Config
 from inmanta.const import LOG_LEVEL_TRACE
 from inmanta.execute.proxy import UnsetException
 from inmanta.execute.runtime import ExecutionContext, ExecutionUnit, QueueScheduler, Resolver
@@ -54,8 +53,8 @@ class Scheduler(object):
         This class schedules statements for execution
     """
 
-    def __init__(self):
-        pass
+    def __init__(self, track_dataflow: bool = False):
+        self.track_dataflow: bool = track_dataflow
 
     def freeze_all(self, exns):
         for t in [t for t in self.types.values() if isinstance(t, Entity)]:
@@ -228,9 +227,9 @@ class Scheduler(object):
         # give all loose blocks an empty XC
         # register the XC's as scopes
         # All named scopes are now present
-        enable_dataflow: bool = Config.get("compiler_datatrace", "enabled", "false") == "true"
+
         for block in blocks:
-            res = Resolver(block.namespace, enable_dataflow)
+            res = Resolver(block.namespace, self.track_dataflow)
             xc = ExecutionContext(block, res)
             block.context = xc
             block.namespace.scope = xc