+++ tests/compiler/test_warnings.py
@@ -36,8 +36,8 @@ def test_warnings(option: Optional[str], expected_error: bool, expected_warning:
     message: str = "Some compiler runtime warning"
     internal_warning: InmantaWarning = CompilerRuntimeWarning(None, message)
     external_warning: Warning = Warning(None, "Some external warning")
+    WarningsManager.apply_config({"default": option} if option is not None else None)
     with warnings.catch_warnings(record=True) as w:
-        WarningsManager.apply_config({"default": option} if option is not None else None)
         if expected_error:
             with pytest.raises(CompilerRuntimeWarning):
                 if raise_external_warning:
@@ -104,6 +104,25 @@ implement A using std::none
         w1 = w[0]
         assert issubclass(w1.category, VariableShadowWarning)
         assert str(w1.message) == message % (2, 8)
+
+
+def test_1918_shadow_warning_for_loop(snippetcompiler):
+    snippetcompiler.setup_for_snippet(
+        """
+i = 0
+
+for i in std::sequence(10):
+end
+        """
+    )
+    message: str = "Variable `i` shadowed: originally declared at {dir}/main.cf:%d, shadowed at {dir}/main.cf:%d"
+    message = message.format(dir=snippetcompiler.project_dir)
+    with warnings.catch_warnings(record=True) as w:
+        compiler.do_compile()
+        assert len(w) == 1
+        w1 = w[0]
+        assert issubclass(w1.category, VariableShadowWarning)
+        assert str(w1.message) == message % (2, 4)
 
 
 def test_deprecation_warning_nullable(snippetcompiler):